working request logging at debugging level
