+++ security/sanitizer.go
@@ -11,8 +11,6 @@
 package security
 
 import (
-	//"html"
-
 	"io"
 	"io/ioutil"
 	"net/http"
@@ -22,14 +20,11 @@ import (
 )
 
 func SanitizeRequest(r *http.Request) {
-
 	if !enabled {
 		return
 	}
-
 	content, _ := ioutil.ReadAll(io.LimitReader(r.Body, 1048576))
 	sanitizedHTML := sanitize.HTML(string(content))
 	r.Body = ioutil.NopCloser(strings.NewReader(sanitizedHTML))
 	r.ContentLength = int64(len(sanitizedHTML))
-
 }
